+++ docs/changelog.md
@@ -1,3 +1,6 @@
+# 2024/12/29 @@石灰
+- feat 完善 `Camera` 类型的加载和 `CameraHelper` 的设置
+
 # 2024/12/28 @@石灰
 - feat 增加场景切换组件 `SceneSwitch` & 部分样式更新
 - feat 调整 `Grid` 中 x 轴 和 z 轴的宽度 调整网格效果